refactor Job type, move to core/schemas
